+++ src/wifi/test/wifi-error-rate-models-test.cc
@@ -66,7 +66,7 @@ WifiErrorRateModelsTestCaseDsss::DoRun (void)
 {
 
   // 1024 bytes plus headers
-  uint32_t size = (1024 + 40 + 14) * 8;
+  uint64_t size = (1024 + 40 + 14) * 8;
   // Spot test some values returned from DsssErrorRateModel
   // Values taken from sample 80211b.c program used in validation paper
   double value;
